requirements.txt
asgiref==3.2.10
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
